✨ 투두 리스트 삭제 확인 기능 구현

+++ src/components/DeleteTodo.js
@@ -2,14 +2,17 @@ import axios from "axios";
 
 function DeleteTodo({ todo, token }) {
   const handleDelete = async () => {
-    await axios.delete(
-      `http://ec2-3-38-135-202.ap-northeast-2.compute.amazonaws.com:8000/todos/${todo.id}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
+    const ok = window.confirm("정말 삭제하시겠습니까?");
+    if (ok) {
+      await axios.delete(
+        `http://ec2-3-38-135-202.ap-northeast-2.compute.amazonaws.com:8000/todos/${todo.id}`,
+        {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        }
+      );
+    }
   };
   return (
     <>
